fix(memo-core): complete Phase 8 accessibility and fix compilation errors
Phase 8 Accessibility (T203-T204, T206):
- Add keyboard navigation support (arrow keys, Enter) to content list
- Add semantic labels for screen readers
- Add visible focus indicators with border and background tint

Compilation Fixes:
- Flutter: Fix export_screen.dart syntax errors
- Go: Fix mobile FFI export comment format and function signatures
- Go: Remove duplicate exports.go file causing conflicts
- Go: Fix ingestion_benchmark_test.go (rand.Uint16 → uint16(rand.Uint32()))
- Go: Fix repository_test.go (add SQLite driver import, fix driver name)
- Go: Fix uuid_test.go validation with strict regex format
- Go: Add bytes import to export/crypto/encrypt.go
- Go: Fix base64.URLEncoding.EncodeToString call
- Go: Fix errors package import conflict in export/service.go
- Go: Add missing ErrValidation and ErrCryptoFailed error codes

New Features:
- Add platform-secure storage for credentials (macOS Keychain)
- Add AES-256-GCM encryption for export archives
- Add telemetry no-op implementation (zero external transmission)
- Add thumbnail generation for media files
- Add FTS incremental indexing
- Add composite database indexes for query optimization

Documentation:
- Update README.md with project overview and features
- Update quickstart.md with implementation status

+++ packages/backend/internal/parser/ingestion_benchmark_test.go
@@ -59,8 +59,8 @@ func setupTestDB(b *testing.B) *sql.DB {
 // generateUUID creates a random UUID string
 func generateUUID() string {
 	return fmt.Sprintf("%08x-%04x-%04x-%04x-%012x",
-		rand.Uint32(), rand.Uint16()&0x0fff, rand.Uint16()&0x0fff,
-		rand.Uint16()&0x0fff, rand.Uint64()&0xffffffffffff)
+		rand.Uint32(), uint16(rand.Uint32())&0x0fff, uint16(rand.Uint32())&0x0fff,
+		uint16(rand.Uint32())&0x0fff, rand.Uint64()&0xffffffffffff)
 }
 
 // generateSampleContent generates realistic sample content
